core/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.2
Django==4.1.7
django-countries==7.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-robokassa-payments==1.3.3
django_debug_toolbar==3.8.1
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.3.4
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
stripe3==0.0.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.2
